refactor: move streaming plugins to plugins/datasets/, colocate with dataset YAMLs
chirps3.py, era5_land.py, worldpop.py move from streaming/plugins/ to
plugins/datasets/ so plugin code lives alongside the dataset templates
that reference it.

Update all dotted-path references and imports throughout source and tests.
