--- open_climate_service/streaming/plugins/__init__.py
@@ -1,1 +1,0 @@
-"""Built-in source plugins for the internal streaming engine."""
