+++ tests/test_dataset_registry.py
@@ -58,7 +58,7 @@ def test_dataset_registry_accepts_supported_sync_kind(
   sync:
     kind: temporal
   ingestion:
-    plugin: open_climate_service.streaming.plugins.chirps3.CHIRPS3DailyPlugin
+    plugin: open_climate_service.plugins.datasets.chirps3.CHIRPS3DailyPlugin
 """,
         encoding="utf-8",
     )
@@ -81,7 +81,7 @@ def test_dataset_registry_accepts_ingestion_plugin_without_function(
   sync:
     kind: temporal
   ingestion:
-    plugin: open_climate_service.streaming.plugins.chirps3.CHIRPS3DailyPlugin
+    plugin: open_climate_service.plugins.datasets.chirps3.CHIRPS3DailyPlugin
 """,
         encoding="utf-8",
     )
@@ -152,7 +152,7 @@ def test_dataset_registry_accepts_supported_sync_execution(
     kind: temporal
     execution: append
   ingestion:
-    plugin: open_climate_service.streaming.plugins.chirps3.CHIRPS3DailyPlugin
+    plugin: open_climate_service.plugins.datasets.chirps3.CHIRPS3DailyPlugin
 """,
         encoding="utf-8",
     )
@@ -177,7 +177,7 @@ def test_dataset_registry_rejects_invalid_sync_availability_function(
     availability:
       latest_available_function: 42
   ingestion:
-    plugin: open_climate_service.streaming.plugins.chirps3.CHIRPS3DailyPlugin
+    plugin: open_climate_service.plugins.datasets.chirps3.CHIRPS3DailyPlugin
 """,
         encoding="utf-8",
     )
@@ -203,7 +203,7 @@ def test_dataset_registry_accepts_sync_availability_function(
     availability:
       latest_available_function: open_climate_service.providers.availability.lagged_latest_available
   ingestion:
-    plugin: open_climate_service.streaming.plugins.chirps3.CHIRPS3DailyPlugin
+    plugin: open_climate_service.plugins.datasets.chirps3.CHIRPS3DailyPlugin
 """,
         encoding="utf-8",
     )
