+++ tests/test_datasets.py
@@ -19,7 +19,7 @@ from open_climate_service.ingestions.schemas import (
     PublicationStatus,
 )
 from open_climate_service.publications.services import managed_dataset_id_for
-from open_climate_service.streaming.plugins.chirps3 import CHIRPS3DailyPlugin
+from open_climate_service.plugins.datasets.chirps3 import CHIRPS3DailyPlugin
 
 
 @pytest.fixture(autouse=True)
@@ -419,7 +419,7 @@ def test_create_artifact_uses_streaming_plugin_for_direct_ingest(
         "variable": "precip",
         "period_type": "daily",
         "ingestion": {
-            "plugin": "open_climate_service.streaming.plugins.chirps3.CHIRPS3DailyPlugin",
+            "plugin": "open_climate_service.plugins.datasets.chirps3.CHIRPS3DailyPlugin",
             "default_params": {"stage": "final"},
         },
     }
@@ -475,7 +475,7 @@ def test_create_artifact_uses_streaming_plugin_for_direct_ingest(
         publish=False,
     )
 
-    assert captured["plugin_path"] == "open_climate_service.streaming.plugins.chirps3.CHIRPS3DailyPlugin"
+    assert captured["plugin_path"] == "open_climate_service.plugins.datasets.chirps3.CHIRPS3DailyPlugin"
     assert captured["params"] == {"stage": "final"}
     assert captured["dataset_id"] == "chirps3_precipitation_daily"
     assert captured["zarr_path"] is None
@@ -509,7 +509,7 @@ def test_create_artifact_uses_streaming_plugin_for_store_based_sync(
         "variable": "precip",
         "period_type": "daily",
         "ingestion": {
-            "plugin": "open_climate_service.streaming.plugins.chirps3.CHIRPS3DailyPlugin",
+            "plugin": "open_climate_service.plugins.datasets.chirps3.CHIRPS3DailyPlugin",
             "default_params": {"stage": "final"},
         },
     }
@@ -576,7 +576,7 @@ def test_create_artifact_forwards_country_code_to_streaming_plugin(
         "variable": "pop_total",
         "period_type": "yearly",
         "ingestion": {
-            "plugin": "open_climate_service.streaming.plugins.worldpop.WorldPopYearlyPlugin",
+            "plugin": "open_climate_service.plugins.datasets.worldpop.WorldPopYearlyPlugin",
             "default_params": {"version": "global2"},
         },
     }
@@ -619,7 +619,7 @@ def test_create_artifact_forwards_country_code_to_streaming_plugin(
         publish=False,
     )
 
-    assert captured["plugin_path"] == "open_climate_service.streaming.plugins.worldpop.WorldPopYearlyPlugin"
+    assert captured["plugin_path"] == "open_climate_service.plugins.datasets.worldpop.WorldPopYearlyPlugin"
     assert captured["params"] == {"version": "global2", "country_code": "SLE"}
     run_kwargs = captured["run"]
     assert isinstance(run_kwargs, dict)
@@ -657,7 +657,7 @@ def test_load_streaming_plugin_rejects_symbol_outside_plugin_protocol(monkeypatc
 
 def test_load_streaming_plugin_filters_runtime_only_params_for_constructor() -> None:
     plugin = services._load_streaming_plugin(
-        "open_climate_service.streaming.plugins.chirps3.CHIRPS3DailyPlugin",
+        "open_climate_service.plugins.datasets.chirps3.CHIRPS3DailyPlugin",
         params={"stage": "final", "country_code": "SLE"},
     )
 
@@ -676,7 +676,7 @@ def test_create_artifact_allows_streaming_coverage_clamped_to_source_availabilit
         "variable": "precip",
         "period_type": "daily",
         "ingestion": {
-            "plugin": "open_climate_service.streaming.plugins.chirps3.CHIRPS3DailyPlugin",
+            "plugin": "open_climate_service.plugins.datasets.chirps3.CHIRPS3DailyPlugin",
         },
     }
     store_path = tmp_path / "chirps3_precipitation_daily.icechunk"
@@ -738,7 +738,7 @@ def test_create_artifact_rejects_streaming_coverage_with_late_start(
         "variable": "precip",
         "period_type": "daily",
         "ingestion": {
-            "plugin": "open_climate_service.streaming.plugins.chirps3.CHIRPS3DailyPlugin",
+            "plugin": "open_climate_service.plugins.datasets.chirps3.CHIRPS3DailyPlugin",
         },
     }
     store_path = tmp_path / "chirps3_precipitation_daily.icechunk"
@@ -786,7 +786,7 @@ def test_create_artifact_returns_409_when_streaming_plugin_has_no_periods(
         "variable": "precip",
         "period_type": "daily",
         "ingestion": {
-            "plugin": "open_climate_service.streaming.plugins.chirps3.CHIRPS3DailyPlugin",
+            "plugin": "open_climate_service.plugins.datasets.chirps3.CHIRPS3DailyPlugin",
         },
     }
     store_path = tmp_path / "chirps3_precipitation_daily.icechunk"
@@ -822,7 +822,7 @@ def test_create_artifact_overwrite_resets_existing_icechunk_store(
         "variable": "precip",
         "period_type": "daily",
         "ingestion": {
-            "plugin": "open_climate_service.streaming.plugins.chirps3.CHIRPS3DailyPlugin",
+            "plugin": "open_climate_service.plugins.datasets.chirps3.CHIRPS3DailyPlugin",
         },
     }
     store_path = tmp_path / "chirps3_precipitation_daily.icechunk"
